feat(library): Library Renaming
Update library path in demo files
